Ignore pixels with 0.0 value, don't calibrate to energies or add to histograms.
Former-commit-id: a61186a5d5797ba4d745ffab9af426dac565e0a9

+++ STFC/HexitecGigEProject/trunk/HxtProcessingLib/generalhxtgenerator.cpp
@@ -3,7 +3,7 @@
 #include <QDebug>
 
 //GeneralHxtGenerator::GeneralHxtGenerator()
-GeneralHxtGenerator::GeneralHxtGenerator(int frameSize, long long binStart, long long binEnd, long long binWidth)
+GeneralHxtGenerator::GeneralHxtGenerator(int frameSize, long long binStart, long long binEnd, double binWidth)
 {
 //   int frameSize = 6400;
    qDebug() << "GeneralHxtGenerator::GeneralHxtGenerator() called";
