+++ STFC/HexitecGigEProject/trunk/HxtProcessingLib/hxtgenerator.h
@@ -7,7 +7,7 @@ class HxtGenerator : public GeneralHxtGenerator
 {
 public:
 //   HxtGenerator();
-   HxtGenerator(int frameSize, long long binStart, long long binEnd, long long binWidth);
+   HxtGenerator(int frameSize, long long binStart, long long binEnd, double binWidth);
 //   double *getNextPixelEnergy();
    void processEnergies(double *pixelEnergy);
